# swagger/openapi.yaml
openapi: 3.0.0
info:
  title: CityLink
  version: 0.0.1
servers:
  - url: http://localhost:4000/api
    description: Dev server
paths:
...
  /tickets/addTicket:
    post:
      summary: Добавление/Покупка билета
      tags: 
        - Tickets
      operationId: addTicket
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
            description: Токен авторизации
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImViMzhmYzFkLTI4YWQtNDY1Mi05OWU5LWJjNjA5YTBkZjllYSIsImlhdCI6MTcwNTg0MzYzOSwiZXhwIjoxNzA2Mjc1NjM5fQ.Xm-cygFpoVrHMaNoiXamGnDb7_LDbGPubNnDWK6BWow
            
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                id:
                  description: id записанного маршрута
                  type: number
                  example: 1
              required:
                - id
      responses:
        '200':
          description: Успешное добавление билета
          content:
            'application/json':
              schema:
                type: object
                properties:
                  ticket:
                    type: object
                    properties:
                      id:
                        description: id билета
                        type: string
                        example: 6141c719-7b67-4228-bff1-312b07816415
                      recordId:
                        description: id записанного маршрута
                        type: number
                        example: 1
                      passengerId:
                        description: id пользователя
                        type: string
                        example: 3a89d97c-54dc-45a1-9ad5-f797f931bcf5
                      seatNo:
                        description: Номер места в автобусе
                        type: number
                        example: 3
                      updatedAt:
                        description: Дата последнего обновления данных билета
                        type: string
                        example: 2024-01-22T21:11:49.546Z
                      createdAt:
                        description: Дата создания билета
                        type: string
                        example: 2024-01-22T21:11:49.546Z
                        
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500"
  /tickets/getUserTickets:
    get:
      summary: Получение списка купленных билетов пользователя
      tags: 
        - Tickets
      operationId: getUserTickets
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
            description: Токен авторизации
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImViMzhmYzFkLTI4YWQtNDY1Mi05OWU5LWJjNjA5YTBkZjllYSIsImlhdCI6MTcwNTg0MzYzOSwiZXhwIjoxNzA2Mjc1NjM5fQ.Xm-cygFpoVrHMaNoiXamGnDb7_LDbGPubNnDWK6BWow
            
      responses:
        '200':
          description: Успешное получение списка купленных билетов пользователя
          content:
            'application/json':
              schema:
                type: object
                properties:
                  tickets:
                    description: Массив объектов
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: id билета
                          example: b55dc887-d36e-4e1a-8101-3c249a47698a
                        seatNo:
                          type: number
                          description: Место в автобусе
                          example: 3
                        price:
                          type: number
                          description: Цена билета
                          example: 500
                        duration:
                          type: number
                          description: Время в пути
                          example: 5
                        departureCity:
                          type: string
                          description: Город отправления
                          example: Ставрополь
                        destinationCity:
                          type: string
                          description: Город прибытия
                          example: Пятигорск
                        departureDate: 
                          type: string
                          description: Дата отправления
                          example: 2023-12-21T08:00:00.000Z
                        
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500" 
  /tickets/getTickets:
    get:
      summary: Получение списка билетов всех пользователей
      description: Возможно получение на две даты (обязательное условие) и по желанию на определённый id пользователя
      tags: 
        - Tickets
      operationId: getTickets
      parameters:
        - name: firstDate
          in: query
          schema:
            type: string
            description: Начальная дата
            example: 1705622400000(timestamp) 
        - name: secondDate
          in: query
          schema:
            type: string
            description: Конечная дата
            example: 1705708800000(timestamp) 
        - name: passengerId
          in: query
          schema:
            type: string
            description: id пользователя
            example: eb38fc1d-28ad-4652-99e9-bc609a0df9ea
        - name: Authorization
          in: header
          schema:
            type: string
            description: Токен авторизации
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImViMzhmYzFkLTI4YWQtNDY1Mi05OWU5LWJjNjA5YTBkZjllYSIsImlhdCI6MTcwNTg0MzYzOSwiZXhwIjoxNzA2Mjc1NjM5fQ.Xm-cygFpoVrHMaNoiXamGnDb7_LDbGPubNnDWK6BWow
    
      responses:
        '200':
          description: Успешное получение списка билетов
          content:
            'application/json':
              schema:
                type: object
                properties:
                  tickets:
                    description: Массив объектов
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: id билета
                          example: b55dc887-d36e-4e1a-8101-3c249a47698a
                        seatNo:
                          type: number
                          description: Место в автобусе
                          example: 3
                        passengerId:
                          type: string
                          description: id пользователя
                          example: eb38fc1d-28ad-4652-99e9-bc609a0df9ea
                        price:
                          type: number
                          description: Цена билета
                          example: 500
                        departureDate: 
                          type: string
                          description: Дата отправления
                          example: 2023-12-23T20:00:00.000Z
                        duration:
                          type: number
                          description: Время в пути
                          example: 5
                        departureCity:
                          type: string
                          description: Город отправления
                          example: Ставрополь
                        destinationCity:
                          type: string
                          description: Город прибытия
                          example: Пятигорск
                        
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500" 
  /tickets/deleteTicket:
    delete:
      summary: Удаление билета
      tags: 
        - Tickets
      operationId: deleteTicket
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
            description: Токен авторизации
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImViMzhmYzFkLTI4YWQtNDY1Mi05OWU5LWJjNjA5YTBkZjllYSIsImlhdCI6MTcwNTg0MzYzOSwiZXhwIjoxNzA2Mjc1NjM5fQ.Xm-cygFpoVrHMaNoiXamGnDb7_LDbGPubNnDWK6BWow
        - name: id
          in: query
          schema:
            type: string
            description: id билета
            example: b89a0496-5e7b-43f8-af16-f7bf07e87570
    
      responses:
        '200':
          description: Успешное удаление билета
          content:
            'application/json':
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Request completed
                        
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500"
components:
  schemas:
    400: 
      description:  Ошибка в отправляемых данных
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Пример сообщения
    500:
      description:  Ошибка на стороне сервера
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Пример сообщения
...